Refactor the tool to match a Flask project structure
Refactored the repo to match solid (best practices) flask structure.
For an idea on what was used, see here: http://flask.pocoo.org/docs/1.0/tutorial/layout/
in line with ticket: T202229.

Also fixing unit tests for the repo to use pytest for resting
the project and updates to README.md.

Bug: T202229
Change-Id: Ic0948780bb287798c0cbccdc540d2c40d0ce0482