[fix] Users added could have entries for a month assumed no entry
This fixes the failures in Ia8ae9aee5422e2f7be1ab37c7813d559159cf7ae
as it shows that this user had contributions for this month which makes
our previous assumption false. So use a far behind past date 1900-12.

The problem with the date 2017-11 is that it's a time that is possible
for a user that is added to have a contribution, which means it will
break this test.

At least we can guarantee that no one had a contribution at that time
in Wikimedia Gerrit, or an alternative can be set this date to a something
like 1900-12 ;-). But one never knows, this tool could still be alive
till that time making us to have this problem again, so the 1900-12 date
remove this problem forever.

Bug: T219098
Change-Id: Ic58f92df834d8b9d3031e71768ae359c868fae15