[IMPR] Completely hide/ignore contribs with 0 patches per abandon policy
If someone submits "N" patches and abandon all ("N") patches, the tool
will hide such a user everywhere possible. Information about this user
won't be shown on any of the charts, on the dashboard, and monthly ranks
as well.

`N - N = 0` and we don't need such a stat on the dashboard as it reveals
no useful information.

Change-Id: I5426fcd8af7a6bfd623f4c44855a4bd542db0bbe